LayoutTests/canvas/philip/tests/2d.fillStyle.parse.current.changed-expected.txt

< [index] >
2d.fillStyle.parse.current.changed
currentColor is computed when the attribute is set, not when it is painted
References: 2d.colours.parse 2d.currentColor.onset
Actual output:
Expected output:

Passed
